Move code into src directory
