+++ keystone.ts
@@ -9,10 +9,10 @@ You can find all the config options in our docs here: https://keystonejs.com/doc
 import { config } from '@keystone-6/core'
 
 // Look in the schema file for how we define our lists, and how users interact with them through graphql or the Admin UI
-import { lists } from './schema'
+import { lists } from './src/schema'
 
 // Keystone auth is configured separately - check out the basic auth setup we are importing from our auth file.
-import { withAuth, session } from './auth'
+import { withAuth, session } from './src/lib/auth'
 
 export default withAuth(
   // Using the config function helps typescript guide you to the available options.
